(#79) rename something_geo -> something_math
