+++ src/something_texture.hpp
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-#include "./something_geo.hpp"
+#include "./something_math.hpp"
 
 using RGBA32 = uint32_t;
